+++ store/commitment/iavl/tree.go
@@ -3,12 +3,13 @@ package iavl
 import (
 	"fmt"
 
-	dbm "github.com/cosmos/cosmos-db"
 	"github.com/cosmos/iavl"
 	ics23 "github.com/cosmos/ics23/go"
 
 	log "cosmossdk.io/log"
+	"cosmossdk.io/store/v2"
 	"cosmossdk.io/store/v2/commitment"
+	dbm "cosmossdk.io/store/v2/db"
 )
 
 var _ commitment.Tree = (*IavlTree)(nil)
@@ -19,8 +20,8 @@ type IavlTree struct {
 }
 
 // NewIavlTree creates a new IavlTree instance.
-func NewIavlTree(db dbm.DB, logger log.Logger, cfg *Config) *IavlTree {
-	tree := iavl.NewMutableTree(db, cfg.CacheSize, cfg.SkipFastStorageUpgrade, logger)
+func NewIavlTree(db store.RawDB, logger log.Logger, cfg *Config) *IavlTree {
+	tree := iavl.NewMutableTree(dbm.NewWrapper(db), cfg.CacheSize, cfg.SkipFastStorageUpgrade, logger)
 	return &IavlTree{
 		tree: tree,
 	}
@@ -67,12 +68,21 @@ func (t *IavlTree) Commit() ([]byte, uint64, error) {
 
 // GetProof returns a proof for the given key and version.
 func (t *IavlTree) GetProof(version uint64, key []byte) (*ics23.CommitmentProof, error) {
-	imutableTree, err := t.tree.GetImmutable(int64(version))
+	immutableTree, err := t.tree.GetImmutable(int64(version))
 	if err != nil {
-		return nil, err
+		return nil, fmt.Errorf("failed to get immutable tree at version %d: %w", version, err)
 	}
 
-	return imutableTree.GetProof(key)
+	return immutableTree.GetProof(key)
+}
+
+func (t *IavlTree) Get(version uint64, key []byte) ([]byte, error) {
+	immutableTree, err := t.tree.GetImmutable(int64(version))
+	if err != nil {
+		return nil, fmt.Errorf("failed to get immutable tree at version %d: %w", version, err)
+	}
+
+	return immutableTree.Get(key)
 }
 
 // GetLatestVersion returns the latest version of the tree.