+++ store/root/store_test.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"testing"
 
-	dbm "github.com/cosmos/cosmos-db"
 	"github.com/stretchr/testify/suite"
 
 	coreheader "cosmossdk.io/core/header"
@@ -12,6 +11,7 @@ import (
 	"cosmossdk.io/store/v2"
 	"cosmossdk.io/store/v2/commitment"
 	"cosmossdk.io/store/v2/commitment/iavl"
+	dbm "cosmossdk.io/store/v2/db"
 	"cosmossdk.io/store/v2/pruning"
 	"cosmossdk.io/store/v2/storage"
 	"cosmossdk.io/store/v2/storage/sqlite"
@@ -102,6 +102,31 @@ func (s *RootStoreTestSuite) TestQuery() {
 	s.Require().NoError(err)
 	s.Require().NotNil(result.ProofOps)
 	s.Require().Equal([]byte("foo"), result.ProofOps[0].Key)
+}
+
+func (s *RootStoreTestSuite) TestGetFallback() {
+	sc := s.rootStore.GetStateCommitment()
+
+	// create a changeset and commit it to SC ONLY
+	cs := store.NewChangeset()
+	cs.Add(testStoreKey, []byte("foo"), []byte("bar"))
+
+	err := sc.WriteBatch(cs)
+	s.Require().NoError(err)
+
+	ci := sc.WorkingCommitInfo(1)
+	_, err = sc.Commit(ci.Version)
+	s.Require().NoError(err)
+
+	// ensure we can query for the key, which should fallback to SC
+	qResult, err := s.rootStore.Query(testStoreKey, 1, []byte("foo"), false)
+	s.Require().NoError(err)
+	s.Require().Equal([]byte("bar"), qResult.Value)
+
+	// non-existent key
+	qResult, err = s.rootStore.Query(testStoreKey, 1, []byte("non_existent_key"), false)
+	s.Require().NoError(err)
+	s.Require().Nil(qResult.Value)
 }
 
 func (s *RootStoreTestSuite) TestQueryProof() {