+++ x/protocolpool/keeper/msg_server_test.go
@@ -399,9 +399,7 @@ func (suite *KeeperTestSuite) TestWithdrawContinuousFund() {
 				suite.Require().NoError(err)
 
 				// Set ToDistribute
-				toDistribute := sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, math.NewInt(100000)))
-				suite.stakingKeeper.EXPECT().BondDenom(suite.ctx).Return("stake", nil).AnyTimes()
-				err = suite.poolKeeper.SetToDistribute(suite.ctx, toDistribute, suite.poolKeeper.GetAuthority())
+				err = suite.poolKeeper.ToDistribute.Set(suite.ctx, math.NewInt(100000))
 				suite.Require().NoError(err)
 			},
 			recipientAddress: []sdk.AccAddress{recipient},
@@ -471,6 +469,7 @@ func (suite *KeeperTestSuite) TestWithdrawContinuousFund() {
 				err = suite.poolKeeper.RecipientFundDistribution.Set(suite.ctx, recipient, math.ZeroInt())
 				suite.Require().NoError(err)
 				toDistribute := sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, math.NewInt(100000)))
+				suite.mockStreamFunds()
 				err = suite.poolKeeper.SetToDistribute(suite.ctx, toDistribute, suite.poolKeeper.GetAuthority())
 				suite.Require().NoError(err)
 			},
@@ -499,6 +498,7 @@ func (suite *KeeperTestSuite) TestWithdrawContinuousFund() {
 				err = suite.poolKeeper.RecipientFundDistribution.Set(suite.ctx, recipient, math.ZeroInt())
 				suite.Require().NoError(err)
 				toDistribute := sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, math.NewInt(100000)))
+				suite.mockStreamFunds()
 				err = suite.poolKeeper.SetToDistribute(suite.ctx, toDistribute, suite.poolKeeper.GetAuthority())
 				suite.Require().NoError(err)
 			},
@@ -565,6 +565,7 @@ func (suite *KeeperTestSuite) TestWithdrawContinuousFund() {
 				suite.Require().NoError(err)
 
 				toDistribute := sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, math.NewInt(100000)))
+				suite.mockStreamFunds()
 				err = suite.poolKeeper.SetToDistribute(suite.ctx, toDistribute, suite.poolKeeper.GetAuthority())
 				suite.Require().NoError(err)
 			},
@@ -617,6 +618,7 @@ func (suite *KeeperTestSuite) TestCreateContinuousFund() {
 	oneMonthInSeconds := int64(30 * 24 * 60 * 60) // Approximate number of seconds in 1 month
 	expiry := suite.ctx.BlockTime().Add(time.Duration(oneMonthInSeconds) * time.Second)
 	testCases := map[string]struct {
+		preRun    func()
 		input     *types.MsgCreateContinuousFund
 		expErr    bool
 		expErrMsg string
@@ -700,11 +702,39 @@ func (suite *KeeperTestSuite) TestCreateContinuousFund() {
 			},
 			expErr: false,
 		},
+		"total funds percentage > 100": {
+			preRun: func() {
+				percentage, err := math.LegacyNewDecFromStr("0.9")
+				recipient2 := sdk.AccAddress([]byte("recipientAddr2___________________"))
+				suite.Require().NoError(err)
+				cf := types.ContinuousFund{
+					Recipient:  recipient2.String(),
+					Percentage: percentage,
+					Expiry:     &time.Time{},
+				}
+				err = suite.poolKeeper.ContinuousFund.Set(suite.ctx, recipient2, cf)
+				suite.Require().NoError(err)
+				intPercentage := percentage.MulInt64(100)
+				err = suite.poolKeeper.RecipientFundPercentage.Set(suite.ctx, recipient2, intPercentage.TruncateInt())
+				suite.Require().NoError(err)
+			},
+			input: &types.MsgCreateContinuousFund{
+				Authority:  suite.poolKeeper.GetAuthority(),
+				Recipient:  recipientAddr.String(),
+				Percentage: percentage,
+				Expiry:     &expiry,
+			},
+			expErr:    true,
+			expErrMsg: "cannot set continuous fund proposal\ntotal funds percentage exceeds 100\ncurrent total percentage: 90",
+		},
 	}
 
 	for name, tc := range testCases {
 		suite.Run(name, func() {
 			suite.SetupTest()
+			if tc.preRun != nil {
+				tc.preRun()
+			}
 
 			_, err := suite.msgServer.CreateContinuousFund(suite.ctx, tc.input)
 			if tc.expErr {
@@ -792,6 +822,7 @@ func (suite *KeeperTestSuite) TestCancelContinuousFund() {
 
 				// Set ToDistribute
 				toDistribute := sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, math.NewInt(100000)))
+				suite.mockStreamFunds()
 				err = suite.poolKeeper.SetToDistribute(suite.ctx, toDistribute, suite.poolKeeper.GetAuthority())
 				suite.Require().NoError(err)